Final clovis renames
