+++ src/backends-data/Motr/CMakeLists.txt
@@ -8,10 +8,10 @@ set(CMAKE_C_FLAGS "-D_REENTRANT -D_GNU_SOURCE -DM0_INTERNAL='' -DM0_EXTERN=exter
 
 set(CMAKE_CXX_FLAGS "-D_REENTRANT -D_GNU_SOURCE -DM0_INTERNAL='' -DM0_EXTERN=extern -iquote ${MOTR_DIR} -iquote '.' -DM0_SRC_DIR=${MOTR_DIR} -fno-common -Wall -Werror -Wno-attributes -fno-strict-aliasing -fno-omit-frame-pointer -g -O2 -rdynamic")
 
-add_library(esdmclovis SHARED clovis.c)
-target_link_libraries(esdmclovis ${GLIB_LIBRARIES})
+add_library(esdmmotr SHARED client.c)
+target_link_libraries(esdmmotr ${GLIB_LIBRARIES})
 include_directories(${ESDM_INCLUDE_DIRS} ${CMAKE_BINARY_DIR} ${GLIB_INCLUDE_DIRS} ${Jansson_INCLUDE_DIRS} ${MOTR_DIR}/motr/.libs)
 
 add_subdirectory(test)
 
-install(TARGETS esdmclovis LIBRARY DESTINATION lib)
+install(TARGETS esdmmotr LIBRARY DESTINATION lib)
