+++ src/backends-data/Motr/Makefile.bak
@@ -5,10 +5,10 @@ AM_CPPFLAGS = -D_REENTRANT -D_GNU_SOURCE -DM0_INTERNAL='' -DM0_EXTERN=extern -iq
 AM_LDFLAGS = -rdynamic -pthread -Wl,-rpath -Wl,/usr/local/lib
 AM_LIBS = ${MOTR_DIR}/lib/libmotr.so -lm -lpthread -laio -lrt -lyaml -luuid ${MOTR_DIR}/lib/libgf_complete.so -ldl ${MOTR_DIR}/lib/libmotr-ut.so ${MOTR_DIR}/lib/libmotr-net-test.so ${MOTR_DIR}/lib/libmotr-xcode-ff2c.so
 
-HDR = clovis.h
-SRC = clovis.c
-OBJ = clovis.o
-TGT = clovis_test
+HDR = client.h
+SRC = client.c
+OBJ = client.o
+TGT = motr_test
 
 all: ${TGT}
 .PHONY: all
