+++ src/backends-data/Motr/test/CMakeLists.txt
@@ -2,9 +2,9 @@ set(CMAKE_C_FLAGS "-D_REENTRANT -D_GNU_SOURCE -DM0_INTERNAL='' -DM0_EXTERN=exter
 
 set(CMAKE_CXX_FLAGS "${CMAKE_C_FLAGS}")
 
-add_executable(clovis_test test.c)
+add_executable(motr_test test.c)
 
 add_library          (motr SHARED IMPORTED)
 set_target_properties(motr PROPERTIES IMPORTED_LOCATION ${MOTR_DIR}/motr/.libs/libmotr.so)
 
-target_link_libraries(clovis_test ${GLIB_LIBRARIES} esdm esdmclovis motr)
+target_link_libraries(motr_test ${GLIB_LIBRARIES} esdm esdmmotr motr)
